Added support to crop vodeo from top and bottom. by default, extend vidoe to PAL/NTSC lines count with black lines

+++ src/iframe.h
@@ -11,22 +11,20 @@ struct IFrame {
   static constexpr uint8_t default_white_lvl = 255;
 
   struct PixelContainer {
-    PixelContainer(size_t width, size_t heigth)
+    PixelContainer(int32_t width, int32_t heigth)
         : pixels(width * heigth), m_width{width}, m_heigth{heigth} {}
 
     myTwoDimArray<uint8_t> getAccessor() {
-      return myTwoDimArray<uint8_t>{pixels.data(),
-                                    static_cast<int32_t>(m_width),
-                                    static_cast<int32_t>(m_heigth)};
+      return myTwoDimArray<uint8_t>{pixels.data(), m_width, m_heigth};
     }
 
-    size_t width() const { return m_width; }
-    size_t heigth() const { return m_heigth; }
+    int32_t width() const { return m_width; }
+    int32_t heigth() const { return m_heigth; }
 
     std::vector<uint8_t> pixels;
 
   private:
-    size_t m_width, m_heigth;
+    int32_t m_width, m_heigth;
   };
 
   enum {
@@ -38,7 +36,7 @@ struct IFrame {
     WHITE_WIDTH = 4
   };
 
-  IFrame(size_t width, size_t heigth) : m_width{width}, m_heigth{heigth} {}
+  IFrame(int32_t width, int32_t heigth) : m_width{width}, m_heigth{heigth} {}
 
   virtual ~IFrame() {}
 
@@ -46,11 +44,11 @@ struct IFrame {
   render(uint8_t grayLevel = default_gray_lvl,
          uint8_t white_lvl = default_white_lvl) const = 0;
 
-  virtual size_t width() const { return m_width; }
-  virtual size_t heigth() const { return m_heigth; }
+  virtual int32_t width() const { return m_width; }
+  virtual int32_t heigth() const { return m_heigth; }
 
 protected:
-  size_t m_width, m_heigth;
+  int32_t m_width, m_heigth;
 };
 
 #endif // IFRAME_H
